+++ src/main/java/mod/gottsch/forge/treasure2/client/model/entity/PirateChestMimicModel.java
@@ -37,7 +37,7 @@ import net.minecraft.world.entity.Entity;
  *
  * @param <T>
  */
-public class PirateChestMimicModel<T extends Entity> extends EntityModel<T> {
+public class PirateChestMimicModel<T extends Entity> extends MimicModel<T> {
 
 	public static final ModelLayerLocation LAYER_LOCATION = new ModelLayerLocation(new ResourceLocation(Treasure.MODID, "pirate_chest_mimic"), "main");
 	
@@ -135,54 +135,79 @@ public class PirateChestMimicModel<T extends Entity> extends EntityModel<T> {
 	/**
 	 * 
 	 */
-	@Override
-	public void setupAnim(T entity, float limbSwing, float limbSwingAmount, float ageInTicks, float netHeadYaw, float headPitch) {
-		PirateChestMimic mimic = (PirateChestMimic)entity;
-		if (mimic.isActive()) {
-			body.xRot = 0.2618F; // 15 degrees
-			
-			// chomp lid
-			if (mimic.hasTarget()) {
-				bobMouth(lid, 22.5f, 22.5f, ageInTicks);
-			}
-			else {
-				bobMouth(lid, 22.5f, 3f, ageInTicks);
-			}
-			
-			latch.xRot = 2.4870942F;
-			eye1.xRot = -1.003564F;
-			eye2.xRot = eye1.xRot;
-			eye3.xRot = eye1.xRot;
-			tongue.xRot = -0.174533F; // 10
-			
-			bob(body, bodyY, ageInTicks);
-		} else {
-			if (mimic.getAmount() < 1F) {
-				body.xRot = mimic.getAmount() * 0.2618F;
-				lid.xRot = mimic.getAmount() * -0.7854F;
-				latch.xRot = mimic.getAmount() * 2.4870942F;
-				eye1.xRot = mimic.getAmount() * -1.003564F;
-				eye2.xRot = eye1.xRot;
-				eye3.xRot = eye1.xRot;
-				tongue.xRot = mimic.getAmount() * -0.174533F;
-			}
-		}
-	}
+//	@Override
+//	public void setupAnim(T entity, float limbSwing, float limbSwingAmount, float ageInTicks, float netHeadYaw, float headPitch) {
+//		PirateChestMimic mimic = (PirateChestMimic)entity;
+//		if (mimic.isActive()) {
+//			body.xRot = 0.2618F; // 15 degrees
+//
+//			// chomp lid
+//			if (mimic.hasTarget()) {
+//				bobMouth(lid, 22.5f, 22.5f, ageInTicks);
+//			}
+//			else {
+//				bobMouth(lid, 22.5f, 3f, ageInTicks);
+//			}
+//
+//			latch.xRot = 2.4870942F;
+//			eye1.xRot = -1.003564F;
+//			eye2.xRot = eye1.xRot;
+//			eye3.xRot = eye1.xRot;
+//			tongue.xRot = -0.174533F; // 10
+//
+//			bob(body, bodyY, ageInTicks);
+//		} else {
+//			if (mimic.getAmount() < 1F) {
+//				body.xRot = mimic.getAmount() * 0.2618F;
+//				lid.xRot = mimic.getAmount() * -0.7854F;
+//				latch.xRot = mimic.getAmount() * 2.4870942F;
+//				eye1.xRot = mimic.getAmount() * -1.003564F;
+//				eye2.xRot = eye1.xRot;
+//				eye3.xRot = eye1.xRot;
+//				tongue.xRot = mimic.getAmount() * -0.174533F;
+//			}
+//		}
+//	}
 
-	public static void bob(ModelPart part, float originY, float age) {
-		part.y = originY + (Mth.cos(age * 0.25F) * 0.5F + 0.05F);
-	}
-	
-	public static void bobMouth(ModelPart mouth, float originRot, float maxRot, float age) {
-		mouth.xRot = -(degToRad(originRot + Mth.cos(age * 0.25f) * maxRot));
-	}
-	
 	@Override
 	public void renderToBuffer(PoseStack poseStack, VertexConsumer vertexConsumer, int packedLight, int packedOverlay, float red, float green, float blue, float alpha) {
 		body.render(poseStack, vertexConsumer, packedLight, packedOverlay, red, green, blue, alpha);
 	}
-	
-	protected static float degToRad(float degrees) {
-		return degrees * (float)Math.PI / 180 ;
+
+	@Override
+	public void activeAnim(T entity, float limbSwing, float limbSwingAmount, float ageInTicks, float netHeadYaw, float headPitch) {
+		super.activeAnim(entity, limbSwing, limbSwingAmount, ageInTicks, netHeadYaw, headPitch);
+
+		latch.xRot = 2.4870942F;
+		eye1.xRot = -1.003564F;
+		eye2.xRot = eye1.xRot;
+		eye3.xRot = eye1.xRot;
+		tongue.xRot = -0.174533F; // 10
+	}
+
+	@Override
+	public void openAnim(T entity, float limbSwing, float limbSwingAmount, float ageInTicks, float netHeadYaw, float headPitch, float timeSlice) {
+		body.xRot = timeSlice * 0.2618F;
+		lid.xRot = timeSlice * getMaxOpenAngle();
+		latch.xRot = timeSlice * 2.4870942F;
+		eye1.xRot = timeSlice * -1.003564F;
+		eye2.xRot = eye1.xRot;
+		eye3.xRot = eye1.xRot;
+		tongue.xRot = timeSlice * -0.174533F;
+	}
+
+	@Override
+	public ModelPart getBody() {
+		return body;
+	}
+
+	@Override
+	public float getBodyY() {
+		return bodyY;
+	}
+
+	@Override
+	public ModelPart getLid() {
+		return lid;
 	}
 }